make cuik featurizer return graph
